move: Info 폴더 이동 ( CalendarInfo, LocationInfo, TimeSchedule )
