+++ server/edusync/src/main/java/com/codestates/edusync/study/studygroup/entity/Studygroup.java
@@ -1,11 +1,11 @@
 package com.codestates.edusync.study.studygroup.entity;
 
 import com.codestates.edusync.audit.Auditable;
-import com.codestates.edusync.study.calendarinfo.entity.CalendarInfo;
+import com.codestates.edusync.infodto.calendarinfo.entity.CalendarInfo;
 import com.codestates.edusync.study.classmate.entity.Classmate;
-import com.codestates.edusync.study.locationinfo.entity.LocationInfo;
+import com.codestates.edusync.infodto.locationinfo.entity.LocationInfo;
 import com.codestates.edusync.study.plancalendar.studygroup.entity.CalendarStudygroup;
-import com.codestates.edusync.study.plancalendar.timeschedule.entity.TimeSchedule;
+import com.codestates.edusync.infodto.timeschedule.entity.TimeSchedule;
 import com.codestates.edusync.study.postcomment.entity.StudyPostComment;
 import com.codestates.edusync.study.studyjoin.entity.StudygroupJoin;
 import com.codestates.edusync.searchtag.entity.SearchTag;
